requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
asn1crypto==1.5.1
astroid==2.9.3
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
colour==0.1.5
crypto==1.4.1
cryptocode==0.1
cryptography==37.0.2
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
dj-static==0.0.6
Django==3.2.9
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-jsonify==0.3.0
django-jsonview==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
httpie==3.0.0
idna==3.3
importlib-metadata==4.10.1
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.0
Markdown==3.3.6
mccabe==0.6.1
multidict==5.2.0
mysqlclient==2.1.0
Naked==0.1.31
oscrypto==1.3.0
Pillow==9.1.0
platformdirs==2.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.14.1
Pygments==2.11.2
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.2
PyPDF3==1.0.6
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.10
requests==2.27.1
requests-toolbelt==0.9.1
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
svglib==1.3.0
swapper==1.3.0
tinycss2==1.1.1
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typed-ast==1.5.3
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
virtualenv==20.13.1
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.7
zipp==3.7.0
